split up the tests.py file and pep8 the tests.
